+++ test/hooks.js
@@ -6,19 +6,24 @@ const packager = require('..')
 const test = require('ava')
 const util = require('./_util')
 
-async function hookTest (wantHookCalled, hookName, t, opts) {
+async function hookTest (wantHookCalled, hookName, t, opts, validator) {
   let hookCalled = false
   opts.dir = util.fixtureSubdir('basic')
   opts.electronVersion = config.version
   opts.arch = 'ia32'
   opts.platform = 'all'
 
-  opts[hookName] = [(buildPath, electronVersion, platform, arch, callback) => {
-    hookCalled = true
-    t.is(electronVersion, opts.electronVersion, `${hookName} electronVersion should be the same as the options object`)
-    t.is(arch, opts.arch, `${hookName} arch should be the same as the options object`)
-    callback()
-  }]
+  opts[hookName] = [validator
+    ? (...args) => {
+      hookCalled = true
+      validator(t, ...args)
+    }
+    : (buildPath, electronVersion, platform, arch, callback) => {
+      hookCalled = true
+      t.is(electronVersion, opts.electronVersion, `${hookName} electronVersion should be the same as the options object`)
+      t.is(arch, opts.arch, `${hookName} arch should be the same as the options object`)
+      callback()
+    }]
 
   // 2 packages will be built during this test
   const finalPaths = await packager(opts)
@@ -28,12 +33,20 @@ async function hookTest (wantHookCalled, hookName, t, opts) {
   t.deepEqual(exists, [true, true], 'Packages should be generated for both 32-bit platforms')
 }
 
-function createHookTest (hookName) {
-  return util.packagerTest(async (t, opts) => hookTest(true, hookName, t, opts))
+function createHookTest (hookName, validator) {
+  return util.packagerTest(async (t, opts) => hookTest(true, hookName, t, opts, validator))
 }
 
 test.serial('platform=all (one arch) for beforeCopy hook', createHookTest('beforeCopy'))
 test.serial('platform=all (one arch) for afterCopy hook', createHookTest('afterCopy'))
+test.serial('platform=all (one arch) for afterFinalizePackageTargets hook', createHookTest('afterFinalizePackageTargets', (t, targets, callback) => {
+  t.is(targets.length, 2, 'target list should have two items')
+  t.is(targets[0].arch, 'ia32')
+  t.is(targets[0].platform, 'linux')
+  t.is(targets[1].arch, 'ia32')
+  t.is(targets[1].platform, 'win32')
+  callback()
+}))
 test.serial('platform=all (one arch) for afterPrune hook', createHookTest('afterPrune'))
 test.serial('platform=all (one arch) for afterExtract hook', createHookTest('afterExtract'))
 test.serial('platform=all (one arch) for afterComplete hook', createHookTest('afterComplete'))