+++ NEWS.md
@@ -7,6 +7,7 @@
 ### Changed
 
 * Replaced `electron-notarize` with `@electron/notarize`.  The configuration options are unchanged. This migration is purely cosmetic.
+* Added new `afterFinalizePackageTargets` hook. This hook exposes the platform/arch combinations that are being packaged when the `arch:all` or `platform:all` options are set.
 
 ## [17.0.0]